+++ examples/filter_module/.steps/01_scaffold.sh
@@ -1,0 +1,2 @@
+just-makeit new my_filters
+cd my_filters