+++ examples/filter_module/.steps/02_module.sh
@@ -1,0 +1,1 @@
+just-makeit module filter